+++ src/CardNumberInput.tsx
@@ -1,0 +1,67 @@
+import React, { useRef } from 'react';
+
+import type { CardNumbersState } from '@/types/types';
+
+interface CardNumberInputProps {
+  cardNumbersStateBundle: [CardNumbersState, React.Dispatch<React.SetStateAction<CardNumbersState>>];
+}
+
+function CardNumberInput({ cardNumbersStateBundle }: CardNumberInputProps) {
+  const [cardNumbers, setCardNumbers] = cardNumbersStateBundle;
+  const cardNumberInputsRef = useRef<(HTMLInputElement | null)[]>([]);
+
+  return (
+    <div className="input-container">
+      <span className="input-title">카드 번호</span>
+      <div className="input-box">
+        {cardNumbers.map(({ key, type, value }, i) => {
+          const isNotLast = i < cardNumbers.length - 1;
+          const isOverThousand = value && value > 1000;
+          const dashComponentClassName = isOverThousand && isNotLast ? 'dash' : 'dash hide';
+
+          if (isNotLast && isOverThousand) {
+            if (document.activeElement === cardNumberInputsRef.current[i]) {
+              cardNumberInputsRef.current[i + 1]?.focus();
+            }
+          }
+
+          return (
+            <>
+              <input
+                key={key}
+                type={type ?? 'text'}
+                value={value ?? ''}
+                className="input-basic text-black"
+                ref={(inputRef) => {
+                  cardNumberInputsRef.current[i] = inputRef;
+                }}
+                onChange={(e) => {
+                  const inputValue = e.currentTarget.value;
+                  const numberValue = inputValue.replace(/\D/g, '');
+                  const fourString = numberValue.substring(0, 4);
+                  setCardNumbers((prev) => {
+                    const inputNumber = fourString ? Number(fourString) : undefined;
+                    if (prev[i].value === inputNumber) {
+                      return prev;
+                    }
+
+                    const newVal = [...prev];
+                    newVal[i] = {
+                      type,
+                      key,
+                      value: inputNumber,
+                    };
+                    return newVal;
+                  });
+                }}
+              />
+              {isNotLast && <div className={`text-black ${dashComponentClassName}`}>-</div>}
+            </>
+          );
+        })}
+      </div>
+    </div>
+  );
+}
+
+export { CardNumberInput };
